requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-storage